Line item update fixes (#18)
* Rename item_id to id

When updating line items we are referencing the unit, not the product. Therefore the proper attribute name is id. That's the same value which is given when creating an order.

* Add line item ID as a required field

To identify line items an ID has to be given to OneStock. Otherwise they will assign a randomly generated value.